requirements.txt
faiss-cpu==1.13.1
numpy>=1.23
pandas>=1.5
google-genai>=0.3.0
youtube-transcript-api>=0.6.1
fastapi>=0.110
uvicorn[standard]>=0.27